remove unnecessary examples
