+++ fltk/examples/README.md
@@ -6,7 +6,7 @@ To run the examples:
 ```
 $ cargo run --example editor
 $ cargo run --example calculator
-$ cargo run --example gallery
+$ cargo run --example tabs
 $ cargo run --example terminal
 $ cargo run --example counter
 $ cargo run --example hello
