+++ fltk/examples/hello.rs
@@ -3,8 +3,8 @@ use fltk::{app::*, frame::*, image::*, window::*};
 fn main() {
     let app = App::default().with_scheme(Scheme::Gleam);
     let mut wind = Window::new(100, 100, 400, 300, "Hello from rust");
-    let mut frame = Frame::new(0, 0, 400, 300, "");
-
+    let mut frame = Frame::new(20, 20, 360, 260, "");
+    frame.set_frame(FrameType::EngraveBox);
     let mut image = SvgImage::load("screenshots/RustLogo.svg").unwrap();
     image.scale(200, 200, true, true);
     frame.set_image(Some(image));
